Refactor #36 - seperate package
