+++ app/src/main/java/com/nlab/reminder/domain/feature/home/di/HomeNavigationModule.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.domain.internal.di.feature.home
+package com.nlab.reminder.domain.feature.home.di
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.setFragmentResultListener
