+++ app/src/main/java/com/nlab/reminder/internal/common/android/database/StringValuesConverter.kt
@@ -14,17 +14,14 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.domain.internal.di.feature.home
+package com.nlab.reminder.internal.common.android.database
 
-import com.nlab.reminder.domain.common.tag.Tag
-import com.nlab.reminder.domain.feature.home.ModifyTagNameUseCase
+import androidx.room.TypeConverter
 
 /**
  * @author Doohyun
  */
-@Deprecated(message = "Fake UseCase was used")
-class FakeModifyTagNameUseCase : ModifyTagNameUseCase {
-    override suspend fun invoke(originalTag: Tag, newText: String) {
-        println("todo implement ModifyTagNameUseCase $originalTag $newText")
-    }
+class StringValuesConverter {
+    @TypeConverter fun listToString(values: List<String>): String = values.joinToString(separator = ",")
+    @TypeConverter fun stringToList(list: String): List<String> = list.split(",")
 }
