+++ app/src/main/res/navigation/nav_main.xml
@@ -16,83 +16,18 @@
 
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
 	xmlns:app="http://schemas.android.com/apk/res-auto"
-	xmlns:tools="http://schemas.android.com/tools"
 	android:id="@+id/home_graph"
-	app:startDestination="@id/homeFragment">
+	app:startDestination="@id/nav_home">
 
 	<action
 		android:id="@+id/action_global_allEndFragment"
-		app:destination="@id/allEndFragment"
+		app:destination="@id/nav_all_end"
 		app:enterAnim="@anim/end_enter"
 		app:exitAnim="@anim/end_exit"
 		app:popEnterAnim="@anim/end_enter_pop"
-		app:popExitAnim="@anim/end_exit_pop"/>
+		app:popExitAnim="@anim/end_exit_pop" />
 
-	<fragment
-		android:id="@+id/homeFragment"
-		android:name="com.nlab.reminder.domain.feature.home.view.HomeFragment"
-		android:label="Home"
-		tools:layout="@layout/fragment_home">
-		<action
-			android:id="@+id/action_homeFragment_to_homeConfigDialogFragment"
-			app:destination="@id/homeTagConfigDialogFragment" />
-		<action
-			android:id="@+id/action_homeFragment_to_homeTagRenameDialogFragment"
-			app:destination="@id/homeTagRenameDialogFragment" />
-		<action
-			android:id="@+id/action_homeFragment_to_homeTagDeleteDialogFragment"
-			app:destination="@id/homeTagDeleteDialogFragment" />
-	</fragment>
-
-	<dialog
-		android:id="@+id/homeTagConfigDialogFragment"
-		android:name="com.nlab.reminder.domain.feature.home.tag.config.view.HomeTagConfigDialogFragment"
-		android:label="HomeTagConfigDialog"
-		tools:layout="@layout/fragment_home_tag_config_dialog">
-		<argument
-			android:name="requestKey"
-			app:argType="string"
-			app:nullable="false" />
-		<argument
-			android:name="tag"
-			app:argType="com.nlab.reminder.domain.common.tag.Tag"
-			app:nullable="false" />
-	</dialog>
-
-	<dialog
-		android:id="@+id/homeTagRenameDialogFragment"
-		android:name="com.nlab.reminder.domain.feature.home.tag.rename.view.HomeTagRenameDialogFragment"
-		android:label="HomeTagRenameDialog"
-		tools:layout="@layout/fragment_home_tag_rename_dialog">
-		<argument
-			android:name="requestKey"
-			app:argType="string"
-			app:nullable="false" />
-		<argument
-			android:name="tag"
-			app:argType="com.nlab.reminder.domain.common.tag.Tag"
-			app:nullable="false" />
-	</dialog>
-
-	<dialog
-		android:id="@+id/homeTagDeleteDialogFragment"
-		android:name="com.nlab.reminder.domain.feature.home.tag.delete.view.HomeTagDeleteDialogFragment"
-		android:label="HomeTagDeleteDialog"
-		tools:layout="@layout/fragment_home_tag_delete_dialog">
-		<argument
-			android:name="requestKey"
-			app:argType="string"
-			app:nullable="false" />
-		<argument
-			android:name="tag"
-			app:argType="com.nlab.reminder.domain.common.tag.Tag"
-			app:nullable="false" />
-	</dialog>
-
-	<fragment
-		android:id="@+id/allEndFragment"
-		android:name="com.nlab.reminder.domain.feature.end.all.view.AllEndFragment"
-		android:label="AllEndFragment"
-		tools:layout="@layout/fragment_all_end" />
+	<include app:graph="@navigation/nav_all_end" />
+	<include app:graph="@navigation/nav_home" />
 
 </navigation>
